Software/Backend/requirements.txt
keras
cvxopt
beautifulsoup4
requests
scikit-learn
sqlalchemy
Flask
Jinja2
flask-restplus
Flask-SQLAlchemy
jsonschema
six
swagger-py-codegen
PyJWT
flask_cors
pandas
numpy
scipy
Cython
jqdatasdk
tushare
tensorflow